Update search bar to include ingredient menu
