+++ client/src/components/SearchBar/Searchbar.css
@@ -1,26 +1,59 @@
 .search-bar {
-  width: 90%;
+  width: 95%;
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+  align-items: center;
 }
 
 .search-button {
   border-radius: 50px;
-  margin-left: 3em;
+  margin-top: 2em;
+  margin-left: 2em;
+  height: 4em;
+}
+
+.search-input-container {
+  position: relative;
+  z-index: 1;
 }
 
 #search-input {
-  background-color: aliceblue;
-  border-radius: 50px;
-  height: 2.5em;
-  font-size: 1.2em;
+  border: 1px solid hsl(0, 0%, 80%);
+  border-radius: 4px;
+  background-color: white;
+}
+
+#ingredient-input {
+  width: 100%;
 }
 
 #search-input::placeholder {
-  color: gray;
+  color: hsl(0, 0%, 50%);
   opacity: 1;
 }
 
-#search-input:focus,
-#search-input:hover {
-  border: 2px solid black;
-  background-color: aliceblue;
+#search-input:focus {
+  border-color: #2684ff;
+  box-shadow: 0 0 0 1px #2684ff;
+}
+
+.tabs-container {
+  position: relative;
+  left: 3px;
+  top: 5px;
+}
+
+.tab {
+  border-top: 0em;
+  border-bottom: 2em solid #f0e68c;
+  border-left: 0.75em solid transparent;
+  border-right: 0.75em solid transparent;
+  background-color: transparent;
+  font-size: 0.8em;
+  height: 0;
+}
+
+.tab.selected {
+  border-bottom: 2em solid #ffd700;
 }
